Extract tui code to own file and refactor main to use tui package
